Moved tests closer to source
