+++ frontend/src/store/resources/circular-reference.json
