+++ frontend/src/store/resources/embedded-collection.json
