+++ frontend/src/store/resources/embedded-single-entity.json
