+++ frontend/src/store/resources/linked-collection.json
