+++ frontend/src/store/resources/linked-single-entity.json
