+++ frontend/src/store/resources/reference-to-single-entity.json
