+++ frontend/src/store/resources/templated-link.json
